- Benchmark: add -s/-distributionseed; remove unused RNGs - PerfTest: Fix a div-by-0 in ResultStats

+++ cs/performance/PerfTest/KeyManager.cs
@@ -35,7 +35,7 @@ namespace FASTER.PerfTest
 
             prevTestResult = null;
 
-            var rng = new RandomGenerator((uint)testResult.Inputs.DistributionSeed);
+            var rng = new RandomGenerator(testResult.Inputs.DistributionSeed);
             var zipfSettings = testResult.Inputs.Distribution == Distribution.Uniform
                 ? null
                 : new ZipfSettings
